Backend: replace HTTP/TLS stack
This is a major change. I got rid of my custom net/http and utls forks. The former has become impossible to maintain by myself and the latter is no longer necessary since my changes have already been merged upstream. I've also taken the opportunity to restructure the codebase and simplify the API in this commit.